requirements.txt
boto3==1.13.24
Django==2.2
django-cors-headers==3.3.0
django-environ==0.4.5
django_graphql_jwt==0.3.1
graphene_django==2.10.1
pre-commit==2.5.1
psycopg2==2.8.5
PyGithub==1.51
pyjwt==1.7.1
redis==3.5.2
requests_oauthlib==1.3.0
stripe==2.48.0
